move project to hw01 directory
