+++ hw01/package.json
@@ -5,7 +5,7 @@
   "main": "convertBytesToHuman.js",
   "type": "module",
   "scripts": {
-    "test": "jest"
+    "test": "../node_modules/.bin/jest"
   },
   "repository": {
     "type": "git",
